+++ lib/deimos/active_record_consume/batch_consumption.rb
@@ -29,21 +29,21 @@ module Deimos
           map { |p, k| Deimos::Message.new(p, nil, key: k) }
 
         tags = %W(topic:#{metadata[:topic]})
+        @valid_active_records = []
+        @invalid_batch_records = []
 
         Deimos.instrument('ar_consumer.consume_batch', tags) do
-          valid_upserts = []
-          invalid_upserts = []
           # The entire batch should be treated as one transaction so that if
           # any message fails, the whole thing is rolled back or retried
           # if there is deadlock
           Deimos::Utils::DeadlockRetry.wrap(tags) do
-            valid_upserts, invalid_upserts = if @compacted || self.class.config[:no_keys]
-                                               update_database(compact_messages(messages))
-                                             else
-                                               uncompacted_update(messages)
-                                             end
+            if @compacted || self.class.config[:no_keys]
+              update_database(compact_messages(messages))
+            else
+              uncompacted_update(messages)
+            end
           end
-          post_process(valid_upserts, invalid_upserts)
+          post_process(@valid_active_records, @invalid_batch_records)
         end
       end
 
@@ -117,64 +117,50 @@ module Deimos
       # All messages are split into slices containing only unique keys, and
       # each slice is handles as its own batch.
       # @param messages [Array<Message>] List of messages.
-      # @return [Array<Array<ActiveRecord::Base>, Array<BatchRecord>>]
+      # @return [Void]
       def uncompacted_update(messages)
-        valid_records = []
-        invalid_records = []
-        BatchSlicer.slice(messages).each do |slice|
-          valid, invalid = update_database(slice)
-          valid_records.push(*valid) if valid.any?
-          invalid_records.push(*invalid) if invalid.any?
-        end
-        [valid_records, invalid_records]
+        BatchSlicer.
+          slice(messages).
+          each(&method(:update_database))
       end
 
       # Perform database operations for a group of messages.
       # All messages with payloads are passed to upsert_records.
       # All tombstones messages are passed to remove_records.
       # @param messages [Array<Message>] List of messages.
-      # @return [Array<Array<ActiveRecord::Base>, Array<BatchRecord>>]
+      # @return [Void]
       def update_database(messages)
         # Find all upserted records (i.e. that have a payload) and all
         # deleted record (no payload)
         removed, upserted = messages.partition(&:tombstone?)
 
         max_db_batch_size = self.class.config[:max_db_batch_size]
-        valid_upserts = []
-        invalid_upserts = []
         if upserted.any?
-          valid_upserts, invalid_upserts = if max_db_batch_size
-                                             upserted.each_slice(max_db_batch_size) do |group|
-                                               valid, invalid = upsert_records(group)
-                                               valid_upserts.push(*valid)
-                                               invalid_upserts.push(*invalid)
-                                             end
-                                             valid_upserts.compact!
-                                             invalid_upserts.compact!
-                                           else
-                                             upsert_records(upserted)
-                                           end
-        end
-
-        if removed.any?
           if max_db_batch_size
-            removed.each_slice(max_db_batch_size) { |group| remove_records(group) }
+            upserted.each_slice(max_db_batch_size) { |group| upsert_records(group) }
           else
-            remove_records(removed)
+            upsert_records(upserted)
           end
         end
-        [valid_upserts, invalid_upserts]
+
+        return if removed.empty?
+
+        if max_db_batch_size
+          removed.each_slice(max_db_batch_size) { |group| remove_records(group) }
+        else
+          remove_records(removed)
+        end
       end
 
       # Upsert any non-deleted records
       # @param messages [Array<Message>] List of messages for a group of
       # records to either be updated or inserted.
-      # @return [Array<Array<ActiveRecord::Base>, Array<BatchRecord>]
+      # @return [Void]
       def upsert_records(messages)
         record_list = build_records(messages)
-        invalid = filter_records(record_list)
+        @invalid_batch_records.concat(filter_records(record_list))
 
-        return [[], invalid] if record_list.empty?
+        return if record_list.empty?
 
         key_col_proc = self.method(:key_columns).to_proc
         col_proc = self.method(:columns).to_proc
@@ -184,14 +170,13 @@ module Deimos
                                   col_proc: col_proc,
                                   replace_associations: self.class.replace_associations,
                                   bulk_import_id_generator: self.class.bulk_import_id_generator)
-        [updater.mass_update(record_list), invalid]
+        @valid_active_records.concat(updater.mass_update(record_list))
       end
 
       # @param record_list [BatchRecordList]
       # @return [Array<BatchRecord>]
       def filter_records(record_list)
         record_list.partition!(self.method(:should_consume?).to_proc)
-        updater.mass_update(record_list)
       end
 
       # Process messages prior to saving to database